Convert base price to float because some discounted prices are not int
For example student discount on HTTPS training results in something.20 CZK.

+++ site/app/Training/Applications.php
@@ -507,7 +507,7 @@ class Applications
 		?string $companyTaxId,
 		?string $note,
 		string $source,
-		?int $price = null,
+		?float $price = null,
 		?float $vatRate = null,
 		?float $priceVat = null,
 		?int $discount = null,
